refactor(transform): isolate shaker invocation effects
